+++ resources/orchard-cmds.clj
@@ -1,6 +1,6 @@
 (defn info [ns-name var-name params]
-  (let [s (orchard.info/info (symbol ns-name) (symbol var-name) (eval params))]
-    (tagged-literal
+  (clojure.core/let [s (orchard.info/info (symbol ns-name) (symbol var-name) (eval params))]
+    (clojure.core/tagged-literal
      'repl-tooling/interactive
      {:html '(cond
                (:ns ?state)
@@ -43,9 +43,69 @@
                [:div.title "Nothing found for this var"])
 
       :state s
-      :fns {:info (list 'fn '[_ s var-name]
-                        (list 'orchard.info/info '(symbol (namespace var-name))
-                                                 '(symbol (name var-name))
-                              params))
-            :change-class '(fn [e s]
-                             (assoc s :sel (symbol (:value e))))}})))
+      :fns {:info (clojure.core/list 'clojure.core/fn '[_ s var-name]
+                    (clojure.core/list 'orchard.info/info
+                                       '(clojure.core/symbol (clojure.core/namespace var-name))
+                                       '(clojure.core/symbol (clojure.core/name var-name))
+                                        params))
+            :change-class '(clojure.core/fn [e s]
+                             (clojure.core/assoc s :sel (clojure.core/symbol (:value e))))}})))
+
+(defn find-usages [symbol-name]
+  (let [sym (clojure.core/symbol symbol-name)
+        refs (orchard.xref/fn-refs sym)
+        grouped (clojure.core/group-by #(-> % meta :ns str) refs)]
+    {:html
+     `[:div.rows
+       [:div.title "Occurrences of " ~symbol-name ":"]
+       [:div.space]
+       ~@(for [ns-name (sort (keys grouped))]
+           `[:div.rows
+             [:div.title "In namespace: " ~ns-name]
+             ~@(for [variable (clojure.core/get grouped ns-name)
+                     :let [v (clojure.core/str (clojure.core/symbol variable))]]
+                 [:div [:a {:href "#"
+                            :on-click (clojure.core/list 'fn '[_]
+                                            (clojure.core/list 'editor/run-feature
+                                                  :go-to-var-definition
+                                                  {:namespace "user"
+                                                   :var-name v}))}
+                        v]])
+             [:div.space]])]}))
+
+(defn clojure-docs [ns-name var-name]
+  (clojure.core/let [doc
+                     (orchard.clojuredocs/find-doc ns-name var-name)]
+    {:html '(let [{:keys [doc nodes examples see-alsos ns name arglists]} ?state]
+              (if doc
+                [:div.rows
+                 [:div.title (:fqn ?state)]
+                 [:<> (map (fn [a] [:div {:key a} "(" (:fqn ?state) " " a ")"]) arglists)]
+                 [:div.space]
+                 [:div.pre doc]
+                 [:div.space]
+                 [:div.title (count examples) " example(s)"]
+                 [:<>
+                  (map (fn [ex i]
+                         (if ((:pages ?state) i)
+                           [:div.rows {:key i}
+                            [:div.cols
+                             [:a.chevron.opened {:href "#" :on-click (?close i)}]
+                             [:div.space]
+                             [:a.icon.clipboard
+                              {:on-click (fn [_] (editor/run-callback :on-copy ex))}]
+                             [:div.pre ex]]
+                            [:div.space]]
+                           [:div.rows {:key i}
+                            [:div.cols
+                             [:a.chevron.closed {:href "#" :on-click (?open i)}]
+                             [:div.space]
+                             (->> ex (take 10) (apply str)) "..."]
+                            [:div.space]]))
+                       examples (range))]]
+                [:div.error "No ClojureDoc for the variable " ns-name (:fqn ?state)]))
+     :state (clojure.core/assoc doc
+                                :pages #{0}
+                                :fqn (str ns-name "/" var-name))
+     :fns '{:open (fn [_ s idx] (clojure.core/update s :pages clojure.core/conj idx))
+            :close (fn [_ s idx] (clojure.core/update s :pages clojure.core/disj idx))}}))