+++ src/repl_tooling/commands_to_repl/all_cmds.cljs
@@ -8,7 +8,8 @@
             [repl-tooling.editor-integration.evaluation :as e-eval]
             [repl-tooling.editor-integration.embedded-clojurescript :as embedded]
             [repl-tooling.editor-integration.definition :as definition]
-            [repl-tooling.editor-integration.doc :as doc]))
+            [repl-tooling.editor-integration.doc :as doc]
+            [repl-tooling.editor-integration.commands :as cmds]))
 
 (defn disconnect!
   "Disconnect all REPLs. Indempotent."
@@ -22,7 +23,7 @@
   "Disconnect all REPLs. Indempotent."
   [state]
   (disconnect!)
-  (reset! state nil))
+  (js/setTimeout #(reset! state nil) 100))
 
 (defn eval-range [state {:keys [contents range] :as data} opts function]
   (let [[start] range
@@ -44,7 +45,8 @@
                                [range (helpers/text-in-range contents range)]))))
 
 (defn all [state {:keys [editor-data] :as opts} repl-kind]
-  (p/let [orchard-cmds (orchard/cmds state)]
+  (p/let [orchard-cmds (orchard/cmds state)
+          config-file (-> @state :editor/callbacks :config-file-path)]
     (cond->
      {:evaluate-top-block {:name "Evaluate Top Block"
                            :description "Evaluates top block block on current editor's selection"
@@ -84,7 +86,14 @@
       :go-to-var-definition {:name "Goto VAR definition"
                              :description "Goes to definition of the current variable"
                              :command #(p/let [data (editor-data)]
-                                         (definition/goto-var data state))}}
+                                         (definition/goto-current-var data state))}}
+
+     config-file
+     (assoc :open-config {:name "Open Config File"
+                          :description "Opens the current config file"
+                          :command #(cmds/run-callback! state :open-editor
+                                                        {:file-name config-file
+                                                          :line 0})})
 
      (= :clj repl-kind)
      (assoc
@@ -97,3 +106,11 @@
                          :old-command #(embedded/connect! state opts false)})
 
      :always (merge orchard-cmds))))
+
+(defn fqn-for-var [editor-state]
+  (p/let [{:keys [contents range filename]} (cmds/run-callback! editor-state :editor-data)
+          [range var] (helpers/current-var contents (first range))
+          res (cmds/run-feature! editor-state :eval
+                                 (str "`" var)
+                                 {:ignore true :auto-detect true :aux true})]
+    (assoc res :range range)))