+++ src/repl_tooling/editor_integration/schemas.cljs
@@ -44,9 +44,32 @@
 (def GotoEditorData
   {:file-name s/Str
    :line s/Int
+   (s/optional-key :column) s/Int
    (s/optional-key :contents) s/Str})
 
-(def Callbacks {:on-start-eval (s/=> s/Any EvalData)
+(def Command {:name s/Str
+              (s/optional-key :description) s/Str
+              (s/optional-key :old-command) (s/=> s/Any)
+              :command (s/=> s/Any)})
+
+(def Commands {:evaluate-top-block Command
+               :evaluate-block Command
+               :evaluate-selection Command
+               :run-tests-in-ns Command
+               :run-test-for-var Command
+               :source-for-var Command
+               :disconnect Command
+               :doc-for-var Command
+               :load-file Command
+               :go-to-var-definition Command
+               (s/optional-key :open-config) Command
+               (s/optional-key :break-evaluation) Command
+               (s/optional-key :connect-embedded) Command
+               s/Keyword Command})
+
+(def Callbacks {:config-file-path (s/maybe s/Str)
+                :register-commands (s/=> s/Any Commands)
+                :on-start-eval (s/=> s/Any EvalData)
                 :open-editor (s/=> s/Any GotoEditorData)
                 :on-eval (s/=> s/Any EvalResult)
                 :editor-data (s/=> EditorData)
@@ -56,7 +79,7 @@
                 :get-config (s/=> Config)
                 :prompt (s/=> s/Any {:title s/Str
                                      :message s/Str
-                                     :arguments [{:key s/Symbol :value s/Str}]})
+                                     :arguments [{:key s/Keyword :value s/Str}]})
                 :on-copy (s/=> s/Any s/Str)
                 :on-stdout (s/=> s/Any s/Str)
                 :on-stderr (s/=> s/Any s/Str)
@@ -66,30 +89,33 @@
                                                         :result ReplResult})
                 :on-disconnect (s/=> s/Any)})
 
-(def Commands {:evaluate-top-block s/Any
-               :evaluate-block s/Any
-               :evaluate-selection s/Any
-               :disconnect s/Any
-               :doc-for-var s/Any
-               ; :spec-for-var s/Any
-               :load-file s/Any
-               (s/optional-key :break-evaluation) s/Any
-               (s/optional-key :connect-embedded) s/Any})
+(def ReplKind (s/enum :clj :cljs :joker :bb :clr :clje))
 
-(def ReplKind (s/enum :clj :cljs :joker :bb :clr))
-
+(def PossibleRanges (s/enum :top-block :block :var :selection :ns))
 (def EditorFeatures {:autocomplete s/Any
-                     :eval-and-render s/Any
-                     :eval s/Any
-                     :result-for-renderer js/Promise})
+                     :eval-and-render (s/=> s/Any s/Any s/Any s/Any)
+                     :evaluate-and-render (s/=> s/Any {:text s/Str
+                                                       :range Range
+                                                       (s/optional-key :pass) s/Any})
+                     :eval (s/=> s/Any s/Any s/Any)
+                     :result-for-renderer (s/=> js/Promise)
+                     :go-to-var-definition (s/=> s/Any {:var-name s/Str
+                                                        :namespace s/Str
+                                                        :repl s/Any})
+                     :get-full-var-name (s/=> js/Promise)
+                     :get-code (s/=> js/Promise PossibleRanges)
+                     :repl-for (s/=> s/Any s/Str (s/enum true false :always nil))})
 
 (def EditorState (s/atom {:editor/callbacks Callbacks
                           :editor/features EditorFeatures
                           (s/optional-key :clj/aux) s/Any
                           (s/optional-key :clj/repl) s/Any
                           (s/optional-key :cljs/repl) s/Any
-                          (s/optional-key :repl/info) {:host s/Str
-                                                       :port s/Int
-                                                       :kind ReplKind
-                                                       :kind-name s/Str}
+                          (s/optional-key :repl/info)
+                          {:host s/Str
+                           :port s/Int
+                           :kind ReplKind
+                           :kind-name s/Str
+                           (s/optional-key :cljs/repl-env) s/Any}
+
                           :editor/commands Commands}))