+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 0.5.0
+- Performance improvement while parsing Clojure code
+- Runnable config
+
+## 0.4.4
+- Interactive renderer support
+
 ## 0.4.1
 - Fixed https://github.com/mauricioszabo/atom-chlorine/issues/150
 - Alpha support for nREPL